+++ lib/Parameter.php
@@ -287,8 +287,22 @@ class Parameter extends Node {
                 if (!is_null($out)) $out.=',';
 
                 // If there's no special characters in the string, we'll use the simple
-                // format
-                if (!preg_match('#(?: [\n":;\^,] )#x', $item)) {
+                // format.
+                //
+                // The list of special characters is defined as:
+                //
+                // Any character except CONTROL, DQUOTE, ";", ":", ","
+                //
+                // by the iCalendar spec:
+                // https://tools.ietf.org/html/rfc5545#section-3.1
+                //
+                // And we add ^ to that because of:
+                // https://tools.ietf.org/html/rfc6868
+                //
+                // But we've found that iCal (7.0, shipped with OSX 10.9)
+                // severaly trips on + characters not being quoted, so we
+                // added + as well.
+                if (!preg_match('#(?: [\n":;\^,\+] )#x', $item)) {
                     return $out.$item;
                 } else {
                     // Enclosing in double-quotes, and using RFC6868 for encoding any