+++ tests/VObject/ITip/BrokerDeleteEventTest.php
@@ -12,6 +12,7 @@ VERSION:2.0
 BEGIN:VEVENT
 UID:foobar
 SEQUENCE:1
+SUMMARY:foo
 ORGANIZER;CN=Strunk:mailto:strunk@example.org
 ATTENDEE;CN=One:mailto:one@example.org
 ATTENDEE;CN=Two:mailto:two@example.org
@@ -41,10 +42,11 @@ PRODID:-//Sabre//Sabre VObject $version//EN
 CALSCALE:GREGORIAN
 METHOD:CANCEL
 BEGIN:VEVENT
-SEQUENCE:2
 UID:foobar
-ATTENDEE;CN=One:mailto:one@example.org
+SEQUENCE:2
+SUMMARY:foo
 ORGANIZER;CN=Strunk:mailto:strunk@example.org
+ATTENDEE;CN=One:mailto:one@example.org
 END:VEVENT
 END:VCALENDAR
 ICS
@@ -65,10 +67,11 @@ PRODID:-//Sabre//Sabre VObject $version//EN
 CALSCALE:GREGORIAN
 METHOD:CANCEL
 BEGIN:VEVENT
-SEQUENCE:2
 UID:foobar
-ATTENDEE;CN=Two:mailto:two@example.org
+SEQUENCE:2
+SUMMARY:foo
 ORGANIZER;CN=Strunk:mailto:strunk@example.org
+ATTENDEE;CN=Two:mailto:two@example.org
 END:VEVENT
 END:VCALENDAR
 ICS
@@ -88,6 +91,7 @@ VERSION:2.0
 BEGIN:VEVENT
 UID:foobar
 SEQUENCE:1
+SUMMARY:foo
 ORGANIZER;CN=Strunk:mailto:strunk@example.org
 ATTENDEE;CN=One:mailto:one@example.org
 ATTENDEE;CN=Two:mailto:two@example.org
@@ -118,7 +122,8 @@ CALSCALE:GREGORIAN
 METHOD:REPLY
 BEGIN:VEVENT
 UID:foobar
-SEQUENCE:2
+SEQUENCE:1
+SUMMARY:foo
 ORGANIZER;CN=Strunk:mailto:strunk@example.org
 ATTENDEE;PARTSTAT=DECLINED;CN=One:mailto:one@example.org
 END:VEVENT
@@ -126,6 +131,35 @@ END:VCALENDAR
 ICS
             ),
         );
+
+        $result = $this->parse($oldMessage, $newMessage, $expected, 'mailto:one@example.org');
+
+
+    }
+
+    function testAttendeeDeleteCancelledEvent() {
+
+        $oldMessage = <<<ICS
+BEGIN:VCALENDAR
+VERSION:2.0
+BEGIN:VEVENT
+STATUS:CANCELLED
+UID:foobar
+SEQUENCE:1
+ORGANIZER;CN=Strunk:mailto:strunk@example.org
+ATTENDEE;CN=One:mailto:one@example.org
+ATTENDEE;CN=Two:mailto:two@example.org
+DTSTART:20140716T120000Z
+END:VEVENT
+END:VCALENDAR
+ICS;
+
+
+        $newMessage = null;
+
+        $version = \Sabre\VObject\Version::VERSION;
+
+        $expected = array();
 
         $result = $this->parse($oldMessage, $newMessage, $expected, 'mailto:one@example.org');