+++ tests/VObject/ParameterTest.php
@@ -87,6 +87,20 @@ class ParameterTest extends \PHPUnit_Framework_TestCase {
 
     }
 
+    /**
+     * iCal 7.0 (OSX 10.9) has major issues with the EMAIL property, when the
+     * value contains a plus sign, and it's not quoted.
+     *
+     * So we specifically added support for that.
+     */
+    function testSerializePlusSign() {
+
+        $cal = new Component\VCalendar();
+        $param = new Parameter($cal, 'EMAIL',"user+something@example.org");
+        $this->assertEquals('EMAIL="user+something@example.org"',$param->serialize());
+
+    }
+
     function testIterate() {
 
         $cal = new Component\VCalendar();